courses/developingapps/python/appengine/end/frontend/requirements.txt
Flask==0.12
google-cloud==0.27.0
google-cloud-core==0.27.1
google-cloud-pubsub==0.28.3
google-cloud-datastore==1.3.0
google-cloud-spanner==0.28.0
google-cloud-storage==1.4.0
gunicorn==19.7.1
